requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.18
jinja2>=3.1.4
pydantic>=2.6.0
pydantic-settings>=2.2.0
Pillow>=10.3.0
pillow-heif>=0.15.0
python-docx>=1.1.0
pypdf>=4.1.0
reportlab>=4.1.0
weasyprint>=61.0
markdown>=3.6
openpyxl>=3.1.2
ffmpeg-python>=0.2.0
pydub>=0.25.1
python-dotenv>=1.0.0
slowapi>=0.1.9

# ── Auth / Database ───────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0
alembic>=1.13
asyncpg>=0.29
python-jose[cryptography]>=3.3.0
bcrypt>=4.0
email-validator>=2.0.0
aiosmtplib>=3.0

# ── Payments ─────────────────────────────────────────────────────
stripe>=8.0

# ── Future: Storage ───────────────────────────────────────────────
# boto3>=1.34
